ETag support and error handling
For files served from storage.zeit.de (epub and mobi), the E-Tag used by
the server is just a md5sum.
Unless `--reload` is specified, the script will now compute the md5
hash and download the file only if it was did not match.

This method does not work for the PDF files, therefore the download is
skipped if the file already exists.

If any request fails, the script will now abort and print the HTTP
status code.